refactor: change naming of argument processor
