+++ fileManager/argumentsCL/startArgsCLContainer.js
@@ -1,0 +1,17 @@
+import { changeSetting } from "../settings.js"
+
+export const longFlagList = {
+    username: addUserName,
+};
+
+export const shortFlagList = {
+
+};
+
+function addUserName(name="noname") {
+    if(name) {
+        changeSetting("username", name);
+    } else {
+        throw new Error("Wrong username");
+    }
+}
